refactor: point api consumers to domain clients
